DK-47 Усовершенствование механизма формирования рейтинга
- Флаг is_counted_in_rating в предметах ЗНО (дефолт выкл): в сумму баллов
  рейтинга и в столбцы DOCX входят только отмеченные предметы (напр. без
  отдельных компонентов НМТ).
- Оценки стали дробными: results_zno.points/points_raw и
  rating_entries.total_points → float; вывод через format_grade.
- Калькулятор комплексного балла (среднее/сумма компонентов) в диалоге
  результата ЗНО карточки абитуриента.
- Миграция d4e5f6a7b8c9, seed.py помечает seed-предметы учитываемыми.

+++ Dekanat/dao/result_zno.py
@@ -33,9 +33,9 @@ class ResultZnoDao:
     def upsert(
         id_items_zno: int,
         id_person: int,
-        points: int,
+        points: float,
         session: Session,
-        points_raw: Optional[int] = None,
+        points_raw: Optional[float] = None,
     ) -> ResultZnoModel:
         existing = ResultZnoDao.get_one(id_items_zno, id_person, session)
         if existing is not None:
